인증로직 추ì가

+++ src/main/java/com/example/leapit/application/bookmark/ApplicationBookmarkController.java
@@ -17,10 +17,10 @@ public class ApplicationBookmarkController {
     // 기업 스크랩 등록 application_bookmark
     @PostMapping("/api/company/bookmark")
     public Resp<?> saveApplicationBookmark(@RequestBody ApplicationBookmarkRequest.SaveDTO reqDTO) {
-        // TODO: session 인증코드 필요
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
 
         try {
-            User sessionUser = (User) session.getAttribute("sessionUser");
             ApplicationBookmarkResponse.SaveDTO respDTO = bookmarkService.saveApplicantBookmarkByUserId(reqDTO, sessionUser.getId());
             return Resp.ok(respDTO);
         } catch (RuntimeException e) {
@@ -31,9 +31,8 @@ public class ApplicationBookmarkController {
     // 기업 스크랩 삭제 application_bookmark
     @DeleteMapping("/api/company/bookmark/{id}")
     public Resp<?> deleteApplicationBookmark(@PathVariable("id") Integer applicationId) {
-        // TODO: session 인증코드 필요
-
         User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) throw new RuntimeException("로그인 후 이용");
         bookmarkService.deleteApplicationBookmarkByApplicationId(applicationId, sessionUser.getId());
         return Resp.ok("북마크 삭제");
     }
